+++ src/math/plane.ts
@@ -1,5 +1,5 @@
+import { Point3D } from "../transform/observable-point";
 import { Ray } from "./ray"
-import { Vec3 } from "./vec3"
 
 function approximately(a: number, b: number) {
   return Math.abs(a - b) <= EPSILON * Math.max(1.0, Math.abs(a), Math.abs(b));
@@ -8,10 +8,10 @@ function approximately(a: number, b: number) {
 const EPSILON = 0.000001
 
 export class Plane {
-  private _normal = new Float32Array(3)
+  private _normal = new Point3D()
 
-  constructor(normal: Float32Array, public distance: number) {
-    Vec3.normalize(normal, this._normal)
+  constructor(normal: Point3D, public distance: number) {
+    this._normal = normal.normalize(this._normal)
   }
 
   get normal() {
@@ -19,11 +19,11 @@ export class Plane {
   }
 
   rayCast(ray: Ray) {
-    const vdot = Vec3.dot(ray.direction, this.normal)
+    const vdot = Point3D.dot(ray.direction, this.normal)
     if (approximately(vdot, 0)) {
       return 0
     }
-    const ndot = -Vec3.dot(ray.origin, this.normal) - this.distance
+    const ndot = -Point3D.dot(ray.origin, this.normal) - this.distance
     return ndot / vdot
   }
 }