+++ src/sprite/projection-sprite.ts
@@ -1,12 +1,13 @@
 import { Sprite } from "@pixi/sprite"
 import { Texture, Resource } from "@pixi/core"
 import { settings } from "@pixi/settings"
+import { Matrix4x4 } from "../transform/matrix"
 
 export class ProjectionSprite extends Sprite {
   private _pixelsPerUnit = 100
 
   distanceFromCamera = 0
-  modelViewProjection = new Float32Array(16)
+  modelViewProjection = new Matrix4x4()
 
   constructor(texture?: Texture<Resource>) {
     super(texture)