+++ src/transform/matrix-component.ts
@@ -1,17 +1,14 @@
 import { TransformId } from "./transform-id"
 
-export class MatrixComponent {
+export class MatrixComponent<T> {
   private _id?: number
-  private _array: Float32Array
 
-  constructor(private _parent: TransformId, size: number, private _update: (array: Float32Array) => void) {
-    this._array = new Float32Array(size)
-  }
+  constructor(private _parent: TransformId, private _data: T, private _update: (data: T) => void) { }
 
-  get array() {
+  get data() {
     if (this._id !== this._parent.transformId) {
-      this._update(this._array); this._id = this._parent.transformId
+      this._update(this._data); this._id = this._parent.transformId
     }
-    return this._array
+    return this._data
   }
 }