+++ test/custom-material.test.mjs
@@ -2,7 +2,7 @@ import { expect } from "chai"
 
 describe("Custom material", () => {
 
-  it("should render correctly when extending material", async () => {
+  it("should render correctly when extending material using pixi *.*.*", async () => {
     let render = (renderer, resources) => {
       let vert = `
         attribute vec3 a_Position;
@@ -18,8 +18,8 @@ describe("Custom material", () => {
         }`
       class CustomMaterial extends PIXI3D.Material {
         updateUniforms(mesh, shader) {
-          shader.uniforms.u_Model = mesh.worldTransform.toArray()
-          shader.uniforms.u_ViewProjection = PIXI3D.Camera.main.viewProjection
+          shader.uniforms.u_Model = mesh.worldTransform.array
+          shader.uniforms.u_ViewProjection = PIXI3D.Camera.main.viewProjection.array
           shader.uniforms.u_Color = [255, 0, 255]
         }
         createShader() {
@@ -40,7 +40,7 @@ describe("Custom material", () => {
     })
   })
 
-  it("should render correctly without extending material", async () => {
+  it("should render correctly without extending material using pixi *.*.*", async () => {
     let render = (renderer, resources) => {
       let vert = `
         attribute vec3 a_Position;
@@ -55,8 +55,8 @@ describe("Custom material", () => {
           gl_FragColor = vec4(u_Color, 1.0);
         }`
       let material = PIXI3D.Material.from(vert, frag, (mesh, shader) => {
-        shader.uniforms.u_Model = mesh.worldTransform.toArray()
-        shader.uniforms.u_ViewProjection = PIXI3D.Camera.main.viewProjection
+        shader.uniforms.u_Model = mesh.worldTransform.array
+        shader.uniforms.u_ViewProjection = PIXI3D.Camera.main.viewProjection.array
         shader.uniforms.u_Color = [255, 0, 255]
       })
       let model = PIXI3D.Model.from(resources["assets/teapot/teapot.gltf"].gltf, {